fixes permissions super admin

+++ app/Http/Controllers/Api/UserController.php
@@ -4,44 +4,59 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UserRequest;
-use App\Http\Resources\UserResource;
+use App\Models\Role;
 use App\Models\User;
-use Illuminate\Http\Resources\Json\JsonResource;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
     public function index()
     {
-        $cash = User::orderByDesc('id')->get();
+        $users = User::orderByDesc('id')->get();
 
-        return response()->json($cash);
+        return response()->json($users);
     }
 
-    public function store(UserRequest $request): JsonResource
+    public function show(User $user)
     {
-        $cash = new User();
-        $cash->fill($request->all());
-        $cash->save();
+        $user->role;
 
-        return new UserResource($cash);
+        return $user;
     }
 
-    public function show(User $user): JsonResource
+    public function store(UserRequest $request)
     {
-        return new UserResource($user);
+        $user = new User();
+        $user->fill($request->all());
+        $roles = Role::find($request['role_id']);
+        $user->save();
+        $user->syncRoles($roles);
+
+        return response()->json($user, Response::HTTP_CREATED);
     }
 
-    public function update(UserRequest $request, User $user): JsonResource
+    public function update(UserRequest $request, User $user)
     {
+        $user->syncRoles([]);
         $user->update($request->all());
+        $roles = Role::find($request['role_id']);
+        $user->syncRoles($roles);
 
-        return new UserResource($user);
+        return response()->json($user);
     }
 
-    public function destroy(User $user): JsonResource
+    public function destroy(User $user)
     {
+        if ($user->id == Auth::user()->id) {
+            abort(422, 'No puedes eliminarte a ti mismo');
+        }
+
         $user->delete();
 
-        return new UserResource($user);
+        return response()->json([
+            'success' => true,
+            'message' => __('general.removed-successfully'),
+        ]);
     }
 }
